Add configuration to run complete app in docker
