+++ mocks/proxy/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "proxy",
+    "version": "1.0.0",
+    "main": "index.js",
+    "scripts": {
+        "start": "node ."
+    },
+    "dependencies": {
+        "express": "^4.18.2",
+        "http-proxy-middleware": "^2.0.6"
+    }
+}
