+++ src/frontend/src/lib/request.js
@@ -1,0 +1,8 @@
+import { dev } from '$app/environment';
+import { env } from '$env/dynamic/private';
+
+export function request({ fetch, url, options }) {
+    const backend = dev ? 'http://localhost:5173' : env.BACKEND_URL;
+    console.log(backend);
+    return fetch(`${backend}${url}`, options);
+}
