--- src/frontend/src/routes/+page.js
@@ -1,6 +1,0 @@
-export async function load({ fetch }) {
-    const response = await fetch('/api/recipe');
-    const data = await response.json();
-
-    return data;
-}
