--- src/frontend/src/routes/create/+page.js
@@ -1,5 +1,0 @@
-export async function load({ fetch }) {
-    const response = await fetch('/api/ingredient');
-
-    return await response.json();
-}
