+++ src/frontend/src/routes/create/+page.server.js
@@ -1,0 +1,7 @@
+import { env } from '$env/dynamic/private';
+
+export async function load({ fetch }) {
+    const response = await fetch(`${env.BACKEND_URL}/api/ingredient`);
+
+    return await response.json();
+}
