--- src/frontend/src/routes/view/[id]/+page.js
@@ -1,7 +1,0 @@
-export async function load({ params, fetch }) {
-    const response = await fetch(`/api/recipe/${params.id}`);
-
-    return {
-        recipe: await response.json()
-    };
-}
