+++ src/proxy/kong.yaml
@@ -1,0 +1,23 @@
+_format_version: '3.0'
+_transform: true
+
+services:
+    - host: rezapp-backend
+      name: backend
+      port: 8080
+      protocol: http
+      routes:
+          - name: api-route
+            paths:
+                - /api
+            strip_path: true
+
+    - host: rezapp-frontend
+      name: frontend
+      port: 3000
+      protocol: http
+      routes:
+          - name: frontend-route
+            paths:
+                - /
+            strip_path: false
